Move schema package out of platform
